recipe_bot/requirements.txt
openai>=1.0.0
langfuse>=2.0.0
braintrust>=0.0.90
python-dotenv>=1.0.0
pydantic>=2.0.0